+++ tests/ResponseTest.php
@@ -62,6 +62,14 @@ class ResponseTest extends TestCase
                 '<?xml version="1.0" encoding="utf-8"?><root></root>',
                 Client::FORMAT_XML
             ],
+            [
+                'access_token=begin|end',
+                Client::FORMAT_URLENCODED
+            ],
+            [
+                'some-plain-string',
+                null
+            ],
         ];
     }